+++ packages/api-server/api_server/models/mixins/_ros_time.py
@@ -1,10 +1,27 @@
-from datetime import datetime
-from typing import Any
+from typing import Any, Optional
 
 from tortoise import fields
-from builtin_interfaces.msg import Time as RosTime
-from .ros_time import ros_to_py_datetime, py_to_ros_time
 
+from datetime import datetime, timezone
+from builtin_interfaces.msg import Time as RosTime
+
+
+def ros_to_py_datetime(ros_time: RosTime) -> datetime:
+    '''
+    ros_time is assumed to be utc.
+    '''
+    return datetime.fromtimestamp(ros_time.sec + ros_time.nanosec / 1000000000, timezone.utc)
+
+
+def py_to_ros_time(py_datetime: datetime) -> RosTime:
+    '''
+    ros_time is assumed to be utc.
+    '''
+    utc_timestamp = py_datetime.timestamp() - py_datetime.utcoffset().total_seconds()
+    return RosTime(
+        sec=int(utc_timestamp),
+        nanosec=int((utc_timestamp % 1) * 1000000000)
+    )
 
 class RosTimeField(fields.DatetimeField):
     '''
